+++ freva-client/src/freva_client/__init__.py
@@ -18,5 +18,5 @@ official documentation: https://freva-org.github.io/freva-legacy
 from .auth import authenticate
 from .query import databrowser
 
-__version__ = "2510.1.2"
+__version__ = "2511.0.0"
 __all__ = ["authenticate", "databrowser", "__version__"]