+++ tests/test_authflows.py
@@ -420,7 +420,7 @@ def test_authenticate_manual_failure_code_flow(
     mocker.patch(
         "freva_client.auth.CodeAuthClient._wait_for_port", return_value=True
     )
-    with patch("freva_rest.auth.server_config.oidc_auth_ports", [8080]):
+    with patch("freva_rest.auth.oauth2.server_config.oidc_auth_ports", [8080]):
         with pytest.raises(OSError, match="No free ports"):
             authenticate(host=test_server, force=True)
 
@@ -454,6 +454,22 @@ def test_token_status(test_server: str, auth: Dict[str, str]) -> None:
     assert res2.status_code != 200
 
 
+def test_wrong_token_claims(
+    test_server: str, mocker: MockerFixture, auth: Dict[str, str]
+) -> None:
+    """Test rejection for wrong token claims."""
+    with mocker.patch(
+        "freva_rest.auth.oauth2.token_field_matches", return_value=False
+    ):
+        res = requests.get(
+            f"{test_server}/data-portal/zarr-utils/status",
+            params={"url": "foo.zar"},
+            headers={"Authorization": f"Bearer {auth['access_token']}"},
+        )
+        assert res.status_code == 401
+        assert "token claims" in res.json()["detail"]
+
+
 # ---------------------------
 # Other utilities.
 # ---------------------------
@@ -468,7 +484,7 @@ def test_auth_utils(free_port: int) -> None:
 
 
 def test_request_headers() -> None:
-    from freva_rest.auth import set_request_header
+    from freva_rest.auth.oauth2 import set_request_header
 
     header, data = {"Content-Type": "foo"}, {}
     set_request_header("foo", "bar", data, header)